revert taskAttemptId to long

+++ client-mr/core/src/main/java/org/apache/hadoop/mapred/SortWriteBufferManager.java
@@ -64,7 +64,7 @@ public class SortWriteBufferManager<K, V> {
   private final Counters.Counter mapOutputRecordCounter;
   private long uncompressedDataLen = 0;
   private long compressTime = 0;
-  private final int taskAttemptId;
+  private final long taskAttemptId;
   private final AtomicLong memoryUsedSize = new AtomicLong(0);
   private final int batch;
   private final AtomicLong inSendListBytes = new AtomicLong(0);
