+++ client-mr/core/src/main/java/org/apache/hadoop/mapreduce/RssMRUtils.java
@@ -252,11 +252,11 @@ public class RssMRUtils {
     return rssJobConf.get(key, defaultValue);
   }
 
-  public static long getBlockId(int partitionId, int taskAttemptId, int nextSeqNo) {
+  public static long getBlockId(int partitionId, long taskAttemptId, int nextSeqNo) {
     return LAYOUT.getBlockId(nextSeqNo, partitionId, taskAttemptId);
   }
 
-  public static long getTaskAttemptId(long blockId) {
+  public static int getTaskAttemptId(long blockId) {
     return LAYOUT.getTaskAttemptId(blockId);
   }
