+++ client-tez/src/main/java/org/apache/tez/runtime/library/common/sort/buffer/WriteBufferManager.java
@@ -95,7 +95,7 @@ public class WriteBufferManager<K, V> {
   private final long sendCheckInterval;
   private final long sendCheckTimeout;
   private final int bitmapSplitNum;
-  private final int taskAttemptId;
+  private final long taskAttemptId;
   private final BlockIdLayout blockIdLayout;
   private TezTaskAttemptID tezTaskAttemptID;
   private final RssConf rssConf;
